moved compiler code all to mod
